+++ SixKeeper/app/src/main/java/com/example/sixkeeper/UserInfoModelClass.kt
@@ -6,5 +6,5 @@ class UserInfoModelClass(
         val lastName: String,
         val birthDate: String,
         val email: String,
-        val mobileNumber: Int
+        val mobileNumber: Long
 )
